ENTMQBR-3597 Statefulset recreated on startup
The issue is caused by multiple updates of one secret.
The first creates the secret then the second resulted
in an update of the statefulset that causes the already
started deployment to restart. The idea is to group
secret creation in one go so it won't trigger the
update of statefulset.